pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "pvc-lastfm"
version = "0.1.0"
requires-python = ">=3.9"

dependencies = [
    "discord.py>=2.3.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "APScheduler>=3.10.0",
    "python-dateutil>=2.8.0",
]

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "UP",  # pyupgrade
]
ignore = [
    "E501", # line too long (handled by formatter)
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
